requirements.txt
# Pillow is used to convert WTPs to RGB images
Pillow
